+++ frontend/src/components/SideBar/PageSideBar.tsx
@@ -4,7 +4,7 @@ import { PageSideBarWrapper } from './SideBar.style';
 const SIDE_BAR_OPTIONS: SideBarOptions[] = [
   { key: '내 카페 관리', value: '/admin/manage-cafe' },
   { key: '내 고객 목록', value: '/admin' },
-  { key: '쿠폰 디자인 제작 및 변경', value: '/admin/make-coupon-policy' },
+  { key: '쿠폰 정책 변경', value: '/admin/modify-coupon-policy/1' },
   { key: '스탬프 적립', value: '/admin/earn-stamp' },
   { key: '리워드 사용', value: '/admin/use-reward' },
 ];
