--- frontend/src/pages/Admin/MakeCouponPolicy/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MakeCouponPolicy = () => {
-  return <div>MakeCouponPolicy</div>;
-};
-
-export default MakeCouponPolicy;
